- Created fragment_login_register.xml and FragmentLoginRegister + Altered names on fragments, to suit proper use + Started out with creating functions for the class and fragment
