+++ .idea/misc.xml
@@ -4,6 +4,7 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Users/jacob/StudioProjects/EksamensProjekt2022/app/src/main/res/layout/fragment_login_home.xml" value="0.358695652173913" />
+        <entry key="..\:/Users/jacob/StudioProjects/EksamensProjekt2022/app/src/main/res/layout/fragment_login_register.xml" value="0.358695652173913" />
         <entry key="..\:/Users/jense/AndroidStudioProjects/EksamensProjekt2022/app/src/main/res/layout/activity_admin.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/jense/AndroidStudioProjects/EksamensProjekt2022/app/src/main/res/layout/activity_employee.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/jense/AndroidStudioProjects/EksamensProjekt2022/app/src/main/res/layout/content_admin.xml" value="0.21770833333333334" />
