--- app/src/main/java/com/jmmnt/UI/FragmentLoginNotInUse.java
@@ -1,40 +1,0 @@
-package com.jmmnt.UI;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.navigation.fragment.NavHostFragment;
-
-import com.jmmnt.R;
-import com.jmmnt.databinding.FragmentLoginNotInUseBinding;
-
-public class FragmentLoginNotInUse extends Fragment {
-
-    private FragmentLoginNotInUseBinding binding;
-
-    @Override
-    public View onCreateView(
-            LayoutInflater inflater, ViewGroup container,
-            Bundle savedInstanceState
-    ) {
-
-        binding = FragmentLoginNotInUseBinding.inflate(inflater, container, false);
-        return binding.getRoot();
-
-    }
-
-    public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-    }
-
-    @Override
-    public void onDestroyView() {
-        super.onDestroyView();
-        binding = null;
-    }
-
-}
