+++ app/src/main/res/values/strings.xml
@@ -5,7 +5,7 @@
     <!-- LOGIN FRAGMENTS -->
     <string name="title_activity_login">LOGIN</string>
     <string name="fragment_login_home_label">LoginHome</string>
-    <string name="fragment_login_not_in_use_label">LoginNotInUse</string>
+    <string name="fragment_login_register_label">LoginNotInUse</string>
 
     <!-- ADMIN FRAGMENTS -->
     <string name="title_activity_admin">ADMIN</string>
